Address fourth round of PR review feedback
- Return admin-specific DTOs instead of EF entities in review endpoints
  (AdminBattleReviewResponse, AdminTicketPurchaseReviewResponse)
- Use AdminLeaderboardEntryResponse DTO with Swagger type annotation
- Fix int overflow in SeasonRepository.AddSeasonWithRoundsAsync by
  casting to long before multiplication
- Remove accidentally committed .claude/scheduled_tasks.lock
- Add .gitignore entry for scheduled_tasks.lock

+++ ArenaService.Shared/Repositories/SeasonRepository.cs
@@ -158,7 +158,7 @@ public class SeasonRepository : ISeasonRepository
         int refreshTicketPolicyId
     )
     {
-        long endBlock = startBlock + (roundInterval * roundCount) - 1;
+        long endBlock = startBlock + ((long)roundInterval * roundCount) - 1;
 
         bool isOverlapping = await IsBlockRangeOverlappingAsync(startBlock, endBlock);
         if (isOverlapping)
